Конвертер валюты [#41583 #144096] + Модульные тесты. [#159435 #144127] chanhed to float type

+++ src/ru/job4j/converter/Converter.java
@@ -2,11 +2,11 @@ package ru.job4j.converter;
 
 public class Converter {
 
-    public static int rubleToEuro(int value) {
+    public static float rubleToEuro(float value) {
         return value / 70;
     }
 
-    public static int rubleToDollar(int value) {
+    public static float rubleToDollar(float value) {
         return value / 60;
     }
 
@@ -15,12 +15,12 @@ public class Converter {
         float dollar = Converter.rubleToDollar(180);
         System.out.println("indicated rubles are " + euro + " euro.");
         System.out.println("indicated rubles are " + dollar + " dollars.");
-            int inEu = 140;
-            int expectedEu = 2;
-            int inDol = 180;
-            int expectedDol = 3;
-            int out1 = Converter.rubleToEuro(inEu);
-            int out2 = Converter.rubleToDollar(inDol);
+            float inEu = 140;
+            float expectedEu = 2;
+            float inDol = 180;
+            float expectedDol = 3;
+            float out1 = Converter.rubleToEuro(inEu);
+            float out2 = Converter.rubleToDollar(inDol);
             boolean passed1 = expectedEu == out1;
             boolean passed2 = expectedDol == out2;
             System.out.println("indicated rubles are 2.0 euro Test result : " + passed1);
